pyproject.toml
[project]
name = "oauth2"
version = "0.1.0"
description = "OAuth2 authentication and authorization service"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
    { name = "Sebastien MORAND", email = "seb.morand@gmail.com" }
]

dependencies = [
    "typer>=0.15.0",
    "rich>=13.9.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "cryptography>=44.0.0",
    "PyJWT>=2.10.0",
    "argon2-cffi>=23.1.0",
    "bcrypt>=4.2.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.18",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "mypy>=1.14.0",
    "ruff>=0.9.0",
    "bandit>=1.8.0",
    "pre-commit>=4.0.0",
]

[project.scripts]
oauth2 = "cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
sources = ["src"]
only-include = ["src"]

# Ruff (Linting + Formatting)
[tool.ruff]
target-version = "py313"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented code)
    "PL",     # Pylint
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",    # Line too long (handled by formatter)
    "PLR0913", # Too many arguments
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "PLR2004", "PLC0415", "PLR0911", "PLR0915", "TC001", "TC003", "RUF012"]
"src/cli.py" = ["PLC0415", "B008"]
"src/models/*.py" = ["TCH003"]
"src/tracing.py" = ["ARG002", "TCH003"]
"src/crypto/*.py" = ["TCH001", "TCH003"]
"src/storage/*.py" = ["TCH001", "TCH003"]
"src/services/*.py" = ["TCH001", "TCH003", "ARG002", "PLR2004"]
"src/routes/*.py" = ["TCH001", "TCH003", "ARG002", "B008"]
"src/middleware/*.py" = ["TCH001", "TCH002", "TCH003"]
"src/app.py" = ["TCH001", "TCH003", "PLC0415"]
"src/dependencies.py" = ["TCH001", "TCH003"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# mypy (Type Checking)
[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
show_column_numbers = true

[[tool.mypy.overrides]]
module = ["httpx.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["argon2.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["bcrypt.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["routes.*"]
warn_return_any = false

# pytest
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["-v", "--tb=short", "--strict-markers"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

# Coverage
[tool.coverage.run]
source = ["src"]
branch = true
omit = ["tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 80
show_missing = true
